requirements.txt
python-telegram-bot==20.6
openai
Pillow
python-dotenv
nest_asyncio
apscheduler
requests
git+https://github.com/openai/whisper.git
torch
ffmpeg-python
deep-translator
feedparser
faster-whisper





